Improve component organisation, hover performance
