+++ package.json
@@ -6,6 +6,7 @@
     "@material-ui/core": "^4.11.4",
     "d3-color": "^3.0.1",
     "deck.gl": "^8.5.8",
+    "lodash": "^4.17.21",
     "mapbox-gl": "^1.13.1",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
@@ -22,6 +23,7 @@
     "@types/d3-color": "^3.0.2",
     "@types/geojson": "^7946.0.8",
     "@types/jest": "^27.0.1",
+    "@types/lodash": "^4.14.173",
     "@types/mapbox-gl": "^1.13.2",
     "@types/node": "^16.7.2",
     "@types/react": "^17.0.19",
