--- src/config/types.ts
@@ -1,1 +1,0 @@
-export type BackgroundName = 'satellite' | 'light';
